Converted tests for GT-11069 to enyo 2.6
