+++ test/moonstone/Tooltip/GT-11069-TooltipChangesDirectionSlider/package.json
@@ -1,0 +1,3 @@
+{
+  "main": "GT-11069-TooltipChangesDirectionSlider.js"
+}
